Add files via upload
In this project it was used node.js, nodemon, .json & speakeasy. The algorithm it was validated using postman by getting a post with userID & the secret key. By getting the userID & the secret key in the database It worked perfectly using the app Authentication and getting the token  and validated with the postman.